+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 **Unreleased**
 --------------
 
+- **Fix**: Fix various compatibility issues with Kotlin 2.2+.
+- **Fix**: Fix `ProducerScope` package in RxObservableEmitDetector.
+- Build against lint `32.2.1`.
+- Target Kotlin 2.2 (matches lint 32.*).
+
 0.11.1
 ------